rename useD -> useData, useV -> useValidation
