+++ lib/components/SDropdownSectionDateRangeYear.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useV } from '../composables/V'
+import { useValidation } from '../composables/Validation'
 import { Year } from '../support/DateRange'
 import { maxValue, minValue, required } from '../validation/rules'
 import SInputNumber from './SInputNumber.vue'
@@ -12,7 +12,7 @@ const emit = defineEmits<{
   (e: 'change', value: Year): void
 }>()
 
-const { validation } = useV(() => ({
+const { validation } = useValidation(() => ({
   year: props.preset.year
 }), {
   year: {
