+++ lib/components/SInputBase.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { type Component, computed, unref, useSlots } from 'vue'
-import { type Validatable } from '../composables/V'
+import { type Validatable } from '../composables/Validation'
 import STooltip from './STooltip.vue'
 import IconQuestion from '~icons/ph/question'
